requirements.txt
torch>=2.0.0
torchaudio>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0
librosa>=0.10.0
soundfile>=0.12.1
audiomentations>=0.30.0
resemblyzer>=0.1.1.dev0
TTS>=0.22.0
streamlit>=1.28.0
gradio>=4.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
pandas>=2.0.0
tqdm>=4.65.0
PyYAML>=6.0
ptflops>=0.7.0
omegaconf>=2.3.0
einops>=0.7.0
